RLEdit/requirements.txt
hydra-core==1.3.2
numpy==1.24.4
omegaconf==2.3.0
sentencepiece==0.1.99
torch==2.3.0+cu121
tqdm==4.66.3
transformers==4.46.0
wandb==0.15.8
nltk==3.9.1
scikit-learn==1.5.2
scipy==1.14.1
datasets==3.1.0